requirements.txt
discord.py==1.6.0
pymongo==3.11.4
dnspython==2.1.0
python-dotenv==0.17.1
requests==2.25.1